urex/requirements.txt
kafka-python
urlextract
